+++ libraries/MTP40C/MTP40C.h
@@ -3,7 +3,7 @@
 //    FILE: MTP40C.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2021-08-20
-// VERSION: 0.3.1
+// VERSION: 0.3.2
 // PURPOSE: Arduino library for MTP40C + MTP40D CO2 sensor
 //     URL: https://github.com/RobTillaart/MTP40C
 //
@@ -13,7 +13,7 @@
 #include "Arduino.h"
 
 
-#define MTP40_LIB_VERSION                     (F("0.3.1"))
+#define MTP40_LIB_VERSION                     (F("0.3.2"))
 
 
 #define MTP40_DEFAULT_ADDRESS                 0x64
@@ -21,6 +21,8 @@
 #define MTP40_OK                              0x00
 #define MTP40_INVALID_AIR_PRESSURE            0x01
 #define MTP40_INVALID_GAS_LEVEL               0x02
+#define MTP40_INVALID_CRC                     0x10
+#define MTP40_NO_STREAM                       0x20
 #define MTP40_INVALID_ADDRESS                 0xFF
 #define MTP40_REQUEST_FAILED                  0xFFFF
 
@@ -28,7 +30,10 @@
 class MTP40
 {
 public:
+  MTP40();
   MTP40(Stream * stream);
+
+  void     setStream(Stream * stream);
 
   bool     begin(uint8_t address = MTP40_DEFAULT_ADDRESS);
   bool     isConnected();
@@ -37,12 +42,12 @@ public:
   bool     setAddress(uint8_t address = MTP40_DEFAULT_ADDRESS);
 
   float    getAirPressureReference();
-  bool     setAirPressureReference(float apr);
+  bool     setAirPressureReference(float apr = 1013);  //  1 Atm default
 
   uint16_t getGasConcentration();  // returns PPM
 
-  void     suppressError(bool se) { _suppressError = se; };
-  bool     getSuppressError()     { return _suppressError; };
+  void     suppressError(bool suppress) { _suppressError = suppress; };
+  bool     getSuppressError()           { return _suppressError; };
 
   //  CALIBRATION FUNCTIONS
   //  READ DATASHEET !!
@@ -52,7 +57,7 @@ public:
   bool     openSelfCalibration();
   bool     closeSelfCalibration();
   uint16_t getSelfCalibrationStatus();
-  bool     setSelfCalibrationHours(uint16_t hours);
+  bool     setSelfCalibrationHours(uint16_t hours = 168);
   uint16_t getSelfCalibrationHours();
 
   void     setGenericAddress()  { _useAddress = false; };
@@ -77,7 +82,7 @@ public:
 //  PROTECTED
 //
 protected:
-  Stream * _ser;
+  Stream * _ser           = NULL;
   uint8_t  _buffer[24];   //  should be big enough.
   uint8_t  _address       = MTP40_DEFAULT_ADDRESS;
 
@@ -92,7 +97,8 @@ protected:
   bool     _suppressError = false;
   int      _lastError     = MTP40_OK;
 
-  bool     request(uint8_t *data, uint8_t cmdlen, uint8_t anslen);
+  void     init();
+  bool     request(uint8_t *data, uint8_t commandLength, uint8_t responseLength);
 
   uint16_t CRC(uint8_t *data, uint16_t len);
 
@@ -111,6 +117,7 @@ protected:
 class MTP40C : public MTP40
 {
 public:
+  MTP40C();
   MTP40C(Stream * str);
 };
 
@@ -118,11 +125,12 @@ public:
 class MTP40D : public MTP40
 {
 public:
+  MTP40D();
   MTP40D(Stream * str);
 
   //  TODO
   //  I2C interface
-  // PWM interface
+  //  PWM interface
 };